pyproject.toml
[tool.poetry]
name = "aisync"
version = "0.1.0"
description = "aisync is a backend service for my portfolio which will mainly focus on AI development"
authors = ["loctvl842 <loclepnvx@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.ruff]
line-length = 150

[tool.black]
line-length = 150

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.4"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pydantic-settings = "^2.2.1"
toml = "^0.10.2"
loguru = "^0.7.2"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
redis = "^5.0.4"
langchain = ">=0.2.6"
langchain-openai = "^0.1.6"
langchain-google-genai = "^1.0.3"
langchain-google-vertexai = "^1.0.4"
langfuse = "^2.31.0"
numpy = "^1.26.4"
gpt4all = "^2.7.0"
langchain-anthropic = "^0.1.16"
langchain-community = "^0.2.6"
tokencost = "^0.1.11"
langgraph = "^0.1.8"
pgvector = "^0.3.1"
sentence-transformers = "^3.0.1"
langchain-ollama = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
